Tests iteración 4 y ajustes generales
Se han implementado la mayor parte de los tests pertenecientes a la iteración 4 y algunos ajustes de funcionamiento para las funciones de algunos controladores.

+++ src/main/java/es/uca/tfg/backend/controller/InterestController.java
@@ -1,8 +1,12 @@
 package es.uca.tfg.backend.controller;
 
+import es.uca.tfg.backend.entity.Event;
 import es.uca.tfg.backend.entity.Interest;
 import es.uca.tfg.backend.entity.Interest;
+import es.uca.tfg.backend.entity.User;
+import es.uca.tfg.backend.repository.EventRepository;
 import es.uca.tfg.backend.repository.InterestRepository;
+import es.uca.tfg.backend.repository.UserRepository;
 import es.uca.tfg.backend.rest.InterestDTO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
@@ -16,7 +20,11 @@ import java.util.Optional;
 public class InterestController {
 
     @Autowired
-    InterestRepository _interestRepository;
+    private InterestRepository _interestRepository;
+    @Autowired
+    private UserRepository _userRepository;
+    @Autowired
+    private EventRepository _eventRepository;
 
     @GetMapping("/getAllInterests")
     public List<Interest> getAllInterests() {
@@ -47,6 +55,19 @@ public class InterestController {
 
     @DeleteMapping("/deleteInterest/{id}")
     public void deleteInterest(@PathVariable("id") int iInterestId) {
-        _interestRepository.deleteById(iInterestId);
+
+        Optional<Interest> optionalInterest = _interestRepository.findById(iInterestId);
+        if(optionalInterest.isPresent()) {
+            Interest interest = optionalInterest.get();
+            for(User user: interest.get_setUsers()) {
+                user.get_setInterests().remove(interest);
+                _userRepository.save(user);
+            }
+            for(Event event: _eventRepository.findByInterest(interest)) {
+                event.get_setInterest().remove(interest);
+                _eventRepository.save(event);
+            }
+            _interestRepository.deleteById(iInterestId);
+        }
     }
 }
